New feature
- Previous socket support on ws://url:port/{mastertoken}/ - this moment only is possible to receive message
- Implement P-Queue to support multiple message receive
- Improve performance
- Config option "disableb64" avoid receive on hook file in b64 format, but only filename
- New endpoint "getFile" - POST - { "filename":<filename>}, returning b64 file

+++ modules/whatsapp.js
@@ -11,27 +11,63 @@ const async = require("async");
 const request = require('request');
 const moment = require('moment');
 const mime = require('mime-types');
+const { default: PQueue } = require("p-queue");
+const queue = new PQueue({timeout: 1000, throwOnTimeout: false });
+
+global.WA_CONFIG_ENV = process.cwd() + '/whatsSessions/config.env';
+
+//get config env
+require('dotenv').config({ path: WA_CONFIG_ENV });
+
 global.uaOverride = 'WhatsApp/2.16.352 Mozilla/5.0 (Macintosh; Intel Mac OS X 10_15_1) AppleWebKit/605.1.15 (KHTML, like Gecko) Version/13.0.3 Safari/605.1.15';
 global.WA_CLIENT = {};
+global.WA_SOCKET = null;
+/*
+* Enviroment Values
+*/
+global.WA_INSTANCE = (F.config['instance'] ? F.config['instance'].toString() : "1") ;
+global.WA_LICENCEKEY = "";
+global.WA_MASTERKEY = "";
+global.WA_TOKENKEY = "";
+global.WA_WEBHOOK = "";
+global.WA_ISDOCKER = false;
+global.WA_DISABLEB64 = false;
+
+//if instance equal 1 operate with enviroment variable
+if( WA_INSTANCE == "1") {
+  WA_LICENCEKEY = (process.env.WA_LICENCEKEY ? process.env.WA_LICENCEKEY : "");
+  WA_MASTERKEY = (process.env.WA_MASTERKEY ? process.env.WA_MASTERKEY : "");
+  WA_TOKENKEY = (process.env.WA_MASTERKEY ? process.env.WA_MASTERKEY : "");
+  WA_WEBHOOK = (process.env.WA_WEBHOOK ? process.env.WA_WEBHOOK : "http://127.0.0.1/");
+  WA_ISDOCKER = true;
+  WA_DISABLEB64 = true;
+} else {
+  WA_LICENCEKEY = (F.config["licensekey"] ? F.config["licensekey"] : "");
+  WA_MASTERKEY = (F.config["masterKey"] ? F.config["masterKey"] : "");
+  WA_TOKENKEY = (F.config["token"] ? F.config["token"] : "");
+  WA_WEBHOOK = (F.config["webhook"] ? F.config["webhook"] : "http://127.0.0.1/");
+  if(F.config["disableb64"] == true)
+    WA_DISABLEB64 = true;
+}
+
 global.WA_CONFIG = {
-    useChrome: true,    
-	licenseKey: (F.config['licensekey'].toString() == "[licence]" ? "" : F.config['licensekey'].toString()),
-    deleteSessionDataOnLogout: false,
-    sessionDataPath: "whatsSessions/",
-    sessionId: F.config['instance'].toString(),
-    headless: true,
+  useChrome: true,    
+	licenseKey: WA_LICENCEKEY,
+  deleteSessionDataOnLogout: false,
+  sessionDataPath: "whatsSessions/",
+  sessionId: WA_INSTANCE,
+  headless: true,
 	hostNotificationLang: 'PT_BR',
 	skipUpdateCheck:true,	
-    autoRefresh:true, 
+  autoRefresh:true, 
 	logFile: true,
-    qrTimeout:0,
+  qrTimeout:0,
 	authTimeout:0,
-    killTimer: 6000,
-    blockCrashLogs: true, 
-    bypassCSP: true,
+  killTimer: 6000,
+  blockCrashLogs: true, 
+  bypassCSP: true,
 	killProcessOnBrowserClose: false,
-	maxChats: 1000,
-	maxMessages: 1000
+  disableSpins: true
   };
 
 /*
@@ -41,6 +77,19 @@ function base64Encode(file) {
   var body = fs.readFileSync(file);
   return body.toString('base64');
 };
+
+/*
+* has socket someone oline
+*/
+function hasSocket(){
+  if(WA_SOCKET) {
+    if(WA_SOCKET.online > 0){
+      return true;
+    }
+  }
+  return false;
+}
+
 
 /*
 * WhatsApp API SUPER CLASS
@@ -74,7 +123,7 @@ var SANITIZE_ACK = function(instanceID,data){
   return JSON.stringify({
       ack: [{
         //id: data.id._serialized,
-		id: data.id,
+		    id: data.id,
         chatId: data.id.remote,
         status: (data.ack == 1 ? 'sent' : (data.ack == 2 ? 'delivered' : 'viewed'))
       }],
@@ -91,11 +140,13 @@ var SANITIZE_MSG = function(instanceID,data) {
   if(DEBUG)
 	  console.log(data);
 
+  let fromName = (data.sender.pushname ? data.sender.pushname : (data.sender.formattedName ? data.sender.formattedName : (data.sender.shortName ? data.sender.shortName : data.from.split('@')[0])));
   return JSON.stringify({
     messages: [{ 
       id: data.id,
       body: data.body,
       filelink: data.filelink,
+      mimetype: (data.mimetype ? data.mimetype : null),
       fromMe: false,
       me: data.to,
       self: 0,
@@ -107,12 +158,12 @@ var SANITIZE_MSG = function(instanceID,data) {
       locIslive:  (data.lng ? (data.isLive ? data.isLive : false) : data.isLive),
       loc: (data.loc ? data.loc : (data.comment ? data.comment : data.loc)),
       chatId: data.chat.id,
-      type: data.type,
-      senderName: (data.sender.pushname ? data.sender.pushname : data.sender.formattedName),
+      type: data.type,      
+      senderName: fromName,
 	    senderPic: data.sender.profilePicThumbObj.eurl,
       caption: (data.caption ? data.caption : null),
       quotedMsgBody: (data.quotedMsgObj ? data.quotedMsgObj : null),	  
-      chatName: (data.isGroupMsg ? data.chat.contact.name : (data.sender.pushname ? data.sender.pushname : data.sender.formattedName))
+      chatName: (data.isGroupMsg ? data.chat.contact.name : fromName)
     }],
     instanceId: instanceID
   });
@@ -131,23 +182,38 @@ WHATS_API.prototype.PROCESS_MESSAGE = function(data){
       if (DEBUG)
         console.log(e);    
     }
-  request({
-    method: 'POST',
-    url:  that.WEBHOOK,
-    headers: { 'Content-Type': 'application/json' },
-    body: SANITIZED
-  }, function(err, response, body){
-    if(err){
+
+  if(hasSocket()) {
+    try {
+      WA_SOCKET.send(SANITIZED);
+
+      if (DEBUG)
+        console.log(SANITIZED);
+    } catch(e) {
       ERROR_CATCHER(err);
-    } else {
-      if(response.statusCode != 200){
-        ERROR_CATCHER("Status Code error: "+response.statusCode,response);
-      } else {
-        if (DEBUG)
-			    console.log(SANITIZED);
-      }
     }
-  });
+  } else {
+
+      //send post 
+      request({
+        method: 'POST',
+        url:  that.WEBHOOK,
+        headers: { 'Content-Type': 'application/json' },
+        body: SANITIZED
+      }, function(err, response, body){
+        if(err){
+          ERROR_CATCHER(err);
+        } else {
+          if(response.statusCode != 200){
+            ERROR_CATCHER("Status Code error: "+response.statusCode,response);
+          } else {
+            if (DEBUG)
+              console.log(SANITIZED);
+          }
+        }
+      });
+
+  }
 };
 
 /*
@@ -157,23 +223,40 @@ WHATS_API.prototype.PROCESS_MESSAGE = function(data){
 WHATS_API.prototype.PROCESS_ACK = function(data){
   var that = this;
   var SANITIZED = SANITIZE_ACK(that.INSTANCE,data);
-  request({
-    method: 'POST',
-    url:  that.WEBHOOK,
-    headers: { 'Content-Type': 'application/json' },
-    body: SANITIZED
-  }, function(err, response, body){
-     if(err){
+
+  if(hasSocket()) {
+    try{
+      WA_SOCKET.send(SANITIZED);
+      
+      if (DEBUG)
+        console.log(SANITIZED);
+
+    } catch(e) {
       ERROR_CATCHER(err);
-    } else {
-      if(response.statusCode != 200){
-        ERROR_CATCHER("Status Code WRONG: "+response.statusCode,response);
-      } else {
-        if (DEBUG)
-		    	console.log(SANITIZED);
-      }
     }
-  });
+
+  } else {
+
+      //send post
+      request({
+        method: 'POST',
+        url:  that.WEBHOOK,
+        headers: { 'Content-Type': 'application/json' },
+        body: SANITIZED
+      }, function(err, response, body){
+        if(err){
+          ERROR_CATCHER(err);
+        } else {
+          if(response.statusCode != 200){
+            ERROR_CATCHER("Status Code WRONG: "+response.statusCode,response);
+          } else {
+            if (DEBUG)
+              console.log(SANITIZED);
+          }
+        }
+      });
+
+  }
 };
 
 /*
@@ -196,37 +279,66 @@ WHATS_API.prototype.SETUP = function(CLIENT,WEBHOOK_INPUT,TOKEN_INPUT) {
   that.TOKEN = TOKEN_INPUT;
   that.CONNECTION = CLIENT;
   CLIENT.onMessage(message => {
-    //CRECKING IF MESSAGE HAVE ANY MEDIA TYPE EMBED
-	//console.log(message);
-     if (message.mimetype) {
-      //SAVING MEDIA RECEIVED AND EXPOSE ADDRESS TO WEB
-      const mediaData = openWA.decryptMedia(message,uaOverride).then(function(DECRYPTED_DATA){
-        var filename = `${message.t}.${mime.extension(message.mimetype)}`;		
-        fs.writeFile(process.cwd()+'/public/cdn/'+filename, Buffer.from(DECRYPTED_DATA, 'base64'), 'base64',function(err) {
-          if(err){
-            console.log("#Error on saving file");
-            message['body'] = `data:${message.mimetype};base64,${message['body']}`;
-            that.PROCESS_MESSAGE(message);
-          } else {
-            message['body'] = `data:${message.mimetype};base64,${base64Encode(process.cwd()+'/public/cdn/'+filename)}`;
-            message['filelink'] = F.ip+':'+F.port+'/cdn/'+filename;
-            that.PROCESS_MESSAGE(message);
-			//no store file in server
-			fs.unlink(process.cwd()+'/public/cdn/'+filename,function(err)  {
-				if (err) {
-					//console.error(err)				
-				} 
-			});
-          }
-        });
-		
-      });	  
-    
-    if (DEBUG)
-	     console.log(imageBase64);
-    } else {
-      that.PROCESS_MESSAGE(message);
-    }
+   queue.add(async () => {
+     
+      //CRECKING IF MESSAGE HAVE ANY MEDIA TYPE EMBED
+      //console.log(message);
+      if (message.mimetype) {
+        //SAVING MEDIA RECEIVED AND EXPOSE ADDRESS TO WEB
+        const mediaData = openWA.decryptMedia(message,uaOverride).then(function(DECRYPTED_DATA){
+          var filename = `${message.t}.${mime.extension(message.mimetype)}`;		
+          fs.writeFile(process.cwd()+'/public/cdn/'+filename, Buffer.from(DECRYPTED_DATA, 'base64'), 'base64', function(err) {
+            if(err){
+              console.log("#Error on saving file");
+              message['body'] = `data:${message.mimetype};base64,${message['body']}`;
+              that.PROCESS_MESSAGE(message);           
+            } else {
+              if(!WA_DISABLEB64) {
+                //if mode docker false, continue using old way
+                message['filelink'] = filename;
+
+                console.log()
+
+                var r = fs.readFile(process.cwd() + '/public/cdn/' + filename, { encoding: 'base64' },  (e, data) => {
+                  if(e) { 
+                    console.log("#Error on saving file");
+                    message['body'] = `data:${message.mimetype};base64,${message['body']}`;
+                    that.PROCESS_MESSAGE(message);           
+                  } else {
+                    message['body'] = `data:${message.mimetype};base64,${data}`;
+                    that.PROCESS_MESSAGE(message);
+
+                      //no store file in server
+                    fs.unlink(process.cwd()+'/public/cdn/'+ filename, function(err)  {
+                      if (err) {
+                        //console.error(err)				
+                      } 
+                    });  
+                  }
+
+                });              
+                
+                
+
+            }  else {
+                message['body'] = "";
+                message['filelink'] = filename;
+                that.PROCESS_MESSAGE(message);
+            }      
+
+            }
+          });
+      
+        });	  
+      
+        /*if (DEBUG)
+          console.log(imageBase64); */
+
+      } else {
+        that.PROCESS_MESSAGE(message);
+      }
+
+   })
   });
 
   CLIENT.onAck(ack => {
@@ -244,6 +356,7 @@ WHATS_API.prototype.SET_QRCODE = function(code){
   if(qrCodeManager){
     qrCodeManager.send({ qr: code });
   };
+
   that.QR_CODE = code;
 };
 
@@ -255,7 +368,7 @@ ON('ready', function(){
   * Creating Connection for WhatsApp and expose conection to WA_CLIENT global var
   * Pay attention to instance id configured on /config file
   */
-  WA_CLIENT = new WHATS_API(F.config['instance']);
+  WA_CLIENT = new WHATS_API(WA_INSTANCE);
 
   /*
   * Declare event getter for when qrcode is available from openWA-api
@@ -269,13 +382,18 @@ ON('ready', function(){
   * Finally creating connection and start headless webBrowser
   * Attention to headless param
   */
- //"/whatsSessions/"+F.config['instance'],
   openWA.create(WA_CONFIG).then(function(client){
+
     //EXECUTING MODULE SETUP
     if(qrCodeManager){
       qrCodeManager.send({ connected: true });
     }
-    WA_CLIENT.SETUP(client, F.config['webhook'], F.config['token']);
+
+     //if have socket send
+     if(WA_SOCKET)
+        WA_SOCKET.send({ connected: true });
+
+    WA_CLIENT.SETUP(client, WA_WEBHOOK, WA_TOKENKEY);
   });
 
 });
